[fix] 푸시 탭 시 notification_type 누락 문제 수정
