+++ HilingualPresentation/Sources/Presentation/Common/DeepLink/DeeplinkManager.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import HilingualCore
 
 public final class DeeplinkManager {
 
@@ -37,5 +38,29 @@ public final class DeeplinkManager {
         case .home, .reminderStreak, .reminderWinback, .reminderCustom:
             nav.popToRootViewController(animated: true)
         }
+    }
+}
+
+extension DeeplinkManager {
+    @discardableResult
+    @MainActor
+    public func handlePushTap(userInfo: [AnyHashable: Any]) -> DeeplinkDestination? {
+        guard let link = userInfo["link"] as? String,
+              let url = URL(string: link),
+              let destination = DeeplinkParser.parse(
+                  url: url,
+                  notificationType: userInfo["notification_type"] as? String
+              ) else {
+            return nil
+        }
+
+        if let analytics = destination.pushNotificationAnalytics {
+            AmplitudeManager.shared.send(
+                .clickPushNotification(notificationType: analytics.type, page: analytics.page)
+            )
+        }
+
+        pendingDestination = destination
+        return destination
     }
 }
